Edit outdated examples
