+++ src/2018-10-18-armadillo-sparse-matrix.cpp
@@ -1,7 +1,7 @@
 // -*- mode: C++; c-indent-level: 4; c-basic-offset: 4; indent-tabs-mode: nil; -*-
 /**
  * @title Sparse matrix in Armadillo
- * @author Dirk Eddelbuettel
+ * @author Dirk Eddelbuettel and Binxiang Ni
  * @license GPL (>= 2)
  * @tags armadillo matrix featured sparse
  * @summary This example shows how to create a sparse matrix in Armadillo
@@ -17,13 +17,13 @@
 
 /*** R
 suppressMessages({
-  library(methods) 
+  library(methods)
   library(Matrix)
 })
-i <- c(1,3:8) 
-j <- c(2,9,6:10) 
+i <- c(1,3:8)
+j <- c(2,9,6:10)
 x <- 7 * (1:7)
-A <- sparseMatrix(i, j, x = x) 
+A <- sparseMatrix(i, j, x = x)
 print(A)
 */
 
@@ -39,34 +39,18 @@ print(A)
 using namespace Rcpp ;
 
 // [[Rcpp::export]]
-void convertSparse(S4 mat) {         // slight improvement with two non-nested loops
+void convertSparse(S4 mat) {
 
+    // obtain dim, i, p. x from S4 object
     IntegerVector dims = mat.slot("Dim");
     arma::urowvec i = Rcpp::as<arma::urowvec>(mat.slot("i"));
-    arma::urowvec p = Rcpp::as<arma::urowvec>(mat.slot("p"));     
+    arma::urowvec p = Rcpp::as<arma::urowvec>(mat.slot("p"));
     arma::vec x     = Rcpp::as<arma::vec>(mat.slot("x"));
 
     int nrow = dims[0], ncol = dims[1];
-    arma::sp_mat res(nrow, ncol);
 
-    // create space for values, and copy
-    arma::access::rw(res.values) = arma::memory::acquire_chunked<double>(x.size() + 1);
-    arma::arrayops::copy(arma::access::rwp(res.values), x.begin(), x.size() + 1);
-
-    // create space for row_indices, and copy
-    arma::access::rw(res.row_indices) = arma::memory::acquire_chunked<arma::uword>(i.size() + 1);
-    arma::arrayops::copy(arma::access::rwp(res.row_indices), i.begin(), i.size() + 1);
-    
-    // create space for col_ptrs, and copy 
-    arma::access::rw(res.col_ptrs) = arma::memory::acquire<arma::uword>(p.size() + 2);
-    arma::arrayops::copy(arma::access::rwp(res.col_ptrs), p.begin(), p.size() + 1);
-
-    // important: set the sentinel as well
-    arma::access::rwp(res.col_ptrs)[p.size()+1] = std::numeric_limits<arma::uword>::max();
-    
-    // set the number of non-zero elements
-    arma::access::rw(res.n_nonzero) = x.size();
-
+    // use Armadillo sparse matrix constructor
+    arma::sp_mat res(i, p, x, nrow, ncol);
     Rcout << "SpMat res:\n" << res << std::endl;
 }
 
@@ -79,11 +63,49 @@ void convertSparse(S4 mat) {         // slight improvement with two non-nested l
 convertSparse(A)
 */
 
+
+
 /**
- * Support for sparse matrix is currently still limited in Armadillo,
- * but expected to grow.  Likewise, RcppArmadillo does not yet have
- * converters such as `as<>()` (though the example above does 
- * essentially everything that is needed) and `wrap()` converters.
- * But we expect to add these
- * eventually --- at which point this example will be much simpler.
+ * Now 11 types of sparse matrix are supported to be converted into RcppArmadillo.
+ * You can just pass a certain type of sparse matrix from R to RcppArmadillo.
+ * It will automatically convert the matrix to a `sp_mat` object.
+ * By the way, back in 2012, we used the below way to create a `sp_mat` Armadillo object.
+ * But now `arma::memory::acquire_chunked` is deprecated. Please avoid using it.
  */
+
+#include <RcppArmadillo.h>
+// [[Rcpp::depends(RcppArmadillo)]]
+
+using namespace Rcpp ;
+
+// [[Rcpp::export]]
+void convertSparse(S4 mat) {         // slight improvement with two non-nested loops
+
+    IntegerVector dims = mat.slot("Dim");
+    arma::urowvec i = Rcpp::as<arma::urowvec>(mat.slot("i"));
+    arma::urowvec p = Rcpp::as<arma::urowvec>(mat.slot("p"));
+    arma::vec x     = Rcpp::as<arma::vec>(mat.slot("x"));
+
+    int nrow = dims[0], ncol = dims[1];
+    arma::sp_mat res(nrow, ncol);
+
+    // create space for values, and copy
+    arma::access::rw(res.values) = arma::memory::acquire_chunked<double>(x.size() + 1);
+    arma::arrayops::copy(arma::access::rwp(res.values), x.begin(), x.size() + 1);
+
+    // create space for row_indices, and copy
+    arma::access::rw(res.row_indices) = arma::memory::acquire_chunked<arma::uword>(i.size() + 1);
+    arma::arrayops::copy(arma::access::rwp(res.row_indices), i.begin(), i.size() + 1);
+
+    // create space for col_ptrs, and copy
+    arma::access::rw(res.col_ptrs) = arma::memory::acquire<arma::uword>(p.size() + 2);
+    arma::arrayops::copy(arma::access::rwp(res.col_ptrs), p.begin(), p.size() + 1);
+
+    // important: set the sentinel as well
+    arma::access::rwp(res.col_ptrs)[p.size()+1] = std::numeric_limits<arma::uword>::max();
+
+    // set the number of non-zero elements
+    arma::access::rw(res.n_nonzero) = x.size();
+
+    Rcout << "SpMat res:\n" << res << std::endl;
+}
